+++ Editor/Inspector/MatAndTexUtils/MatAndTexAbsoluteSeparatorEditor.cs
@@ -21,7 +21,7 @@ namespace net.rs64.TexTransTool.Editor.MatAndTexUtils
 
             var sTargetRenderers = thisSObject.FindProperty("TargetRenderers");
             var sMultiRendererMode = thisSObject.FindProperty("MultiRendererMode");
-            TextureTransformerEditor.DrawerRenderer(sTargetRenderers, sMultiRendererMode.boolValue);
+            TextureTransformerEditor.DrawerRenderer(sTargetRenderers, "CommonDecal:prop:TargetRenderer".Glc(), sMultiRendererMode.boolValue);
             EditorGUILayout.PropertyField(sMultiRendererMode);
 
             if (_tempMaterial == null || GUILayout.Button("Refresh Materials")) { RefreshMaterials(sTargetRenderers, ref _tempMaterial); }
@@ -120,7 +120,7 @@ namespace net.rs64.TexTransTool.Editor.MatAndTexUtils
         {
             var sObj = new SerializedObject(target);
             var sTargetRenderers = sObj.FindProperty("TargetRenderers");
-            TextureTransformerEditor.DrawerTargetRenderersSummary(sTargetRenderers);
+            TextureTransformerEditor.DrawerTargetRenderersSummary(sTargetRenderers,sTargetRenderers.name.Glc());
             sObj.ApplyModifiedProperties();
         }
     }